requirements.txt
PyInstaller==6.16.0
Pillow==12.0.0
pdfplumber==0.11.9
playwright==1.55.0
flet>=0.25.0
